picoporj: fix another -Wshorten-64-to-32 warning from clang
  ../examples/picoproj.c:196:8: error: implicit conversion loses integer precision: 'size_t'
        (aka 'unsigned long') to 'int' [-Werror,-Wshorten-64-to-32]
          ret = fread(image, size, 1, image_fp);
              ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ examples/picoproj.c
@@ -193,7 +193,7 @@ int main(int argc, char *argv[])
 		goto out_close_image_fp;
 	}
 
-	ret = fread(image, size, 1, image_fp);
+	ret = (int)fread(image, size, 1, image_fp);
 	if (ret != 1) {
 		if (feof(image_fp))
 			fprintf(stderr, "Unexpected end of file.\n");
